backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
supabase==2.9.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
sse-starlette==2.2.1
httpx==0.27.2
PyJWT==2.9.0
cryptography==43.0.3
